Updating WebEd 4

+++ src/Actions/DeleteUserAction.php
@@ -24,6 +24,14 @@ class DeleteUserAction extends AbstractAction
     {
         $id = do_filter(BASE_FILTER_BEFORE_DELETE, $id, WEBED_USERS);
 
+        if ($id == get_current_logged_user_id()) {
+            return response_with_messages(
+                trans('webed-users::base.cannot_update_status_yourself'),
+                true,
+                \Constants::FORBIDDEN_CODE
+            );
+        }
+
         $result = $this->repository->delete($id, $force);
 
         do_action(BASE_ACTION_AFTER_DELETE, WEBED_USERS, $id, $result);
